Add origin to MySQL result. This allowes us to determine from which table the data comes from.
* fix an issue with va_list, now all tables should load as espected.

+++ src/world/Objects/Units/Players/PlayerDefines.hpp
@@ -832,7 +832,7 @@ static const uint32_t ClassRaceCombinations[91][3] =
 
 inline uint32_t getAEVersion()
 {
-    return BUILD_VERSION;
+    return static_cast<uint32_t>(BUILD_VERSION);
 }
 
 inline bool isClassRaceCombinationPossible(uint8_t _class, uint8_t _race)
